Add abstraction layer ("Client") between Model and Engine (#1122)
This implements an abstraction layer that sits
between the Model class (which manages user interactions with the
guidance DSL) and the engine class (which actually "runs" the program
specified by the DSL against a local model).

This PR introduces two main abstractions:
1. `State`
- Encapsulates a sequence of "chunks" corresponding to inputs (e.g.
text, images, etc.), outputs (e.g. text/tokens, images, audio, etc.
- The chunk sequence is _generic_ and independent of the underlying LM
implementation, so we only need a single visualization/introspection
layer around it
- `State` also contains an _LM-specific_ representation of the chunk
seqence, e.g. formatted as a list of messages in OpenAI format, a string
"prompt" for use with completion models, etc.
- It is up to an implementer to add methods for handling a each chunk
type if they want to add support for a new LM
- We can provide default blanket implementations (e.g. one that works
for most `transformers` models without multi-modal inputs) that can be
easily extended (e.g. adding support for Phi-3-vision should only
require subclassing the base implementation and adding a single method
to handle incoming images)
3. `Client`
- Maintains no internal state -- it is up to the `Model` class to store
the `State` object and make updates to it
    - `Model` class calls `Client`'s `run` function
- accepts a `State` (corresponding to a prompt) and an AST node from the
guidance DSL (corresponding to a request for generation with some
constraints) and yields chunks to be applied to the `State`
- The `Model` class should have a hook that updates the visualization
layer as each chunk is applied
- It is an _implementation detail_ that some `Client` subclasses talk to
the existing guidance `Engine` class (which runs the guidance DSL with
`llguidance`).
- The generic implementation of this is given by the `GuidanceClient`
class
- I currently have a `TransformersClient` subclass that handles
interaction with the existing `guidance.models.transformers` backend,
but I would love to push this code down into a "server-side"
implementation so we only ever need the one `GuidanceClient`
- I have provided a stub implementation of an `OpenAIClient` that talks
to the OpenAI API directly (entirely bypassing the `Engine` class) as a
proof of concept. Note that this PR does NOT attempt to "transpile" the
guidance DSL to a control-loop around OpenAI's API -- that will be
coming in the next PR.


All of this is in the `guidance.experimental` submodule, but to merge
this PR, I want to get to a place where the experimental `Model` class
serves as a drop-in replacement for the existing one.

---------
