+++ guidance/_grammar.py
@@ -1,8 +1,8 @@
 import re
 import types
-from typing import TYPE_CHECKING, Any, Optional, Sequence, Union, cast
+from typing import Any, Optional, Sequence, Union, cast 
+from dataclasses import dataclass
 
-from . import _parser
 
 # to support the embedding of guidance functions inside Python f-strings we use tags with these delimiters
 tag_start = "{{G|"  # start of a call tag
@@ -57,40 +57,29 @@ class RawFunction(Function):
         self.kwargs = kwargs
 
     def __call__(self, model):
-        return self.f(model, *self.args, **self.kwargs)
+        model = self.f(model, *self.args, **self.kwargs)
+        if model is None:
+            raise Exception(
+                f"The guidance function `{self.f.__name__}` did not return a model object! You need to return an updated model object at the end of your guidance function."
+            )
+        return model
 
     def __add__(self, other):
-
-        # if we are joining with a string we use the string representation for ourselves
-        if isinstance(other, str):
-            return str(self) + other
-
+        if not isinstance(other, (str, bytes, Function)):
+            return NotImplemented
         def __add__(model):
             model = self(model)
-            if model is None:
-                raise Exception(
-                    f"The guidance function `{self.f.__name__}` did not return a model object! You need to return an updated model object at the end of your guidance function."
-                )
-            if isinstance(other, GrammarFunction):
-                return model + other
-            else:
-                return other(model)
-
+            model += other
+            return model
         return RawFunction(__add__, [], {})
 
     def __radd__(self, other):
-
-        # if we are joining with a string we use the string representation for ourselves
-        if isinstance(other, str):
-            return other + str(self)
-
+        if not isinstance(other, (str, bytes, Function)):
+            return NotImplemented
         def __radd__(model):
-            if isinstance(other, GrammarFunction):
-                model += other
-            else:
-                model = other(model)
-            return self(model)
-
+            model += other
+            model = self(model)
+            return model
         return RawFunction(__radd__, [], {})
 
 
@@ -142,9 +131,8 @@ class GrammarFunction(Function):
         if isinstance(value, GrammarFunction):
             return Join([self, value])
 
-        # otherwise we let the stateful object handle things
-        else:
-            return value.__radd__(self)
+        # otherwise we let the other object handle things
+        return NotImplemented
 
     def __radd__(self, value):
 
@@ -159,9 +147,8 @@ class GrammarFunction(Function):
         if isinstance(value, GrammarFunction):
             return Join([value, self])
 
-        # otherwise we let the stateful object handle things
-        else:
-            return value.__add__(self)
+        # otherwise we let the other object handle things
+        return NotImplemented
 
     def __getitem__(self, value):
         raise StatefulException("GrammarFunctions can't access state!")
@@ -174,13 +161,15 @@ class GrammarFunction(Function):
     ) -> Union[Match, None]:
         if isinstance(byte_string, str):
             byte_string = byte_string.encode()
-        parser = _parser.ByteParser(self)
+
+        from ._parser import ByteParser, ByteParserException
+        parser = ByteParser(self)
 
         try:
             parser.consume_bytes(byte_string)
             if not allow_partial:
                 parser.force_done()
-        except _parser.ByteParserException:
+        except ByteParserException:
             if raise_exceptions:
                 raise
             else:
@@ -195,7 +184,8 @@ class GrammarFunction(Function):
         return Match(*parser.get_captures(), partial=not parser.matched())  # type: ignore[misc]
 
     def forced_prefix(self) -> str:
-        parser = _parser.ByteParser(self)
+        from ._parser import ByteParser
+        parser = ByteParser(self)
         return parser.bytes.decode("utf-8", errors="ignore")
 
     @classmethod
@@ -656,6 +646,17 @@ class LLGrammar(WithMaxTokens):
     def __init__(self, grammar_with_lexer: dict[str, Any], *, capture_name: Optional[float] = None, temperature: Optional[float] = None, max_tokens: Optional[int] = None):
         super().__init__(capture_name=capture_name, temperature=temperature, max_tokens=max_tokens)
         self.grammar_with_lexer = grammar_with_lexer
+
+
+# These are "abstract" role tags (i.e. they indicate the start and end of a role, but they do not know the corresponding text/tokens).
+# They are "grammar-adjacent" concepts but not really grammars themselves.
+@dataclass
+class RoleStart:
+    role: str
+
+@dataclass
+class RoleEnd:
+    role: str
 
 
 def string(value: Union[str, bytes]) -> Union[Null, Join]:
