+++ guidance/_server.py
@@ -1,7 +1,10 @@
-import base64
 import os
 
 from typing import TYPE_CHECKING, Iterator
+
+from .models import Model
+from .models._engine import Engine
+from ._schema import EngineCallResponse
 
 try:
     import pydantic
@@ -12,9 +15,6 @@ try:
 except ImportError:
     if TYPE_CHECKING:
         raise
-
-from .models._model import Model, Engine
-from ._schema import EngineCallResponse
 
 
 class GuidanceRequest(pydantic.BaseModel):
