+++ guidance/library/_block.py
@@ -1,18 +1,21 @@
-from guidance import models
+from contextlib import contextmanager
+from typing import Optional
+from .._ast import ASTNode
+from ..models._base._model import _active_blocks
 
-
-class ContextBlock:
-    def __init__(self, opener, closer, name=None):
+class Block:
+    def __init__(self, name: Optional[str], opener: Optional[ASTNode], closer: Optional[ASTNode]):
+        self.name = name
         self.opener = opener
         self.closer = closer
-        self.name = name
-
-    def __enter__(self):
-        models.Model.global_active_blocks.append(self)
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        models.Model.global_active_blocks.remove(self)
 
 
-def block(name=None, opener="", closer=""):
-    return ContextBlock(opener, closer, name=name)
+@contextmanager
+def block(name=None, opener=None, closer=None):
+    current_blocks = _active_blocks.get()
+    new_block = Block(name=name, opener=opener, closer=closer)
+    token = _active_blocks.set(current_blocks + (new_block,))
+    try:
+        yield
+    finally:
+        _active_blocks.reset(token)
