+++ guidance/library/_role.py
@@ -1,34 +1,16 @@
-from .._guidance import guidance
-from ._block import block
-from ._set_attribute import set_attribute
-
-
-@guidance
-def role_opener(lm, role_name, **kwargs):
-    lm += lm.role_opener(role_name, **kwargs)
-    return lm
-
-
-@guidance
-def role_closer(lm, role_name, **kwargs):
-    lm += lm.role_closer(role_name, **kwargs)
-    return lm
-
+from ._block import block, Block
+from .._grammar import RoleStart, RoleEnd
 
 # TODO HN: Add a docstring to better describe arbitrary role functions
-def role(role_name, text=None, **kwargs):
-    if text is None:
-        return block(
-            name=role_name,
-            opener=role_opener(role_name, **kwargs),
-            closer=role_closer(role_name, **kwargs),
-        )
-    else:
-        assert False
-        # return self.append(open_text + text + close_text)
+def role(role: str) -> Block:
+    return block(
+        name=None,
+        opener=RoleStart(role),
+        closer=RoleEnd(role),
+    )
 
 
-def system(text=None, **kwargs):
+def system() -> Block:
     """Indicate the 'system' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -41,10 +23,10 @@ def system(text=None, **kwargs):
         >>>     lm += "A system prompt"
 
     """
-    return role("system", text, **kwargs)
+    return role("system")
 
 
-def user(text=None, **kwargs):
+def user() -> Block:
     """Indicate the 'user' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -57,10 +39,10 @@ def user(text=None, **kwargs):
         >>>     lm += "What the user said"
 
     """
-    return role("user", text, **kwargs)
+    return role("user")
 
 
-def assistant(text=None, **kwargs):
+def assistant() -> Block:
     """Indicate the 'assistant' prompt
 
     A convention has grown up around 'chat' APIs that
@@ -74,16 +56,4 @@ def assistant(text=None, **kwargs):
         >>>     lm += gen(name="model_output", max_tokens=20)
 
     """
-    return role("assistant", text, **kwargs)
-
-
-def function(text=None, **kwargs):
-    return role("function", text, **kwargs)
-
-
-def instruction(text=None, **kwargs):
-    return role("instruction", text, **kwargs)
-
-
-def indent_roles(indent=True):
-    return set_attribute("indent_roles", indent)
+    return role("assistant")
