+++ guidance/models/__init__.py
@@ -1,30 +1,31 @@
-from ._model import Model, Instruct, Chat
+from ._base import Model
+# from ._engine import Instruct, Chat
 
 # local models
-from .transformers._transformers import Transformers, TransformersTokenizer
+from .transformers import Transformers, TransformersTokenizer
 from .llama_cpp import LlamaCpp
-from ._mock import Mock, MockChat
+from ._mock import Mock#, MockChat
 
 # grammarless models (we can't do constrained decoding for them)
-from ._grammarless import Grammarless
-from .vertexai._vertexai import (
-    VertexAI,
-    VertexAIChat,
-    VertexAICompletion,
-    VertexAIInstruct,
-)
-from ._azure_openai import (
-    AzureOpenAI,
-)
-from ._azureai_studio import AzureAIStudioChat
+# from ._grammarless import Grammarless
+# from .vertexai._vertexai import (
+#     VertexAI,
+#     VertexAIChat,
+#     VertexAICompletion,
+#     VertexAIInstruct,
+# )
+# from ._azure_openai import (
+#     AzureOpenAI,
+# )
+# from ._azureai_studio import AzureAIStudioChat
 from ._openai import OpenAI
-from ._lite_llm import LiteLLM, LiteLLMChat, LiteLLMInstruct, LiteLLMCompletion
-from ._cohere import Cohere, CohereCompletion, CohereInstruct
-from ._anthropic import Anthropic
-from ._googleai import GoogleAI, GoogleAIChat
-from ._togetherai import (
-    TogetherAI,
-    TogetherAIChat,
-    TogetherAIInstruct,
-    TogetherAICompletion,
-)
+# from ._lite_llm import LiteLLM, LiteLLMChat, LiteLLMInstruct, LiteLLMCompletion
+# from ._cohere import Cohere, CohereCompletion, CohereInstruct
+# from ._anthropic import Anthropic
+# from ._googleai import GoogleAI, GoogleAIChat
+# from ._togetherai import (
+#     TogetherAI,
+#     TogetherAIChat,
+#     TogetherAIInstruct,
+#     TogetherAICompletion,
+# )
