+++ guidance/models/_googleai.py
@@ -1,5 +1,5 @@
 import re
-from ._model import Chat, Instruct
+from ._engine._engine import Chat, Instruct
 from ._grammarless import Grammarless, GrammarlessEngine
 import tiktoken
 import os
