+++ guidance/models/_lite_llm.py
@@ -1,6 +1,6 @@
 import tiktoken
 
-from ._model import Chat, Instruct
+from ._engine._engine import Chat, Instruct
 from ._grammarless import GrammarlessTokenizer, GrammarlessEngine, Grammarless
