+++ guidance/models/transformers/_engine.py
@@ -2,16 +2,16 @@ import os
 import re
 import textwrap
 import warnings
+from typing import Any, Optional, Sequence, Union
 
-from typing import Sequence, Union
-
-from guidance._schema import GenToken, GenTokenExtra
+from ..._schema import GenToken, GenTokenExtra
+from .._base import Message
+from .._engine import Engine, Tokenizer
 
 try:
     import torch
 except ModuleNotFoundError:
     pass
-
 
 try:
     import transformers as transformers_package
@@ -20,9 +20,6 @@ try:
 except ModuleNotFoundError:
     has_transformers = False
 
-
-from .._model import Engine, Model
-from .._tokenizer import Tokenizer
 
 # Formed by comparing model and tokenizer from_pretrained methods
 # transformers/models/auto/auto_factory.py
@@ -384,15 +381,17 @@ class TransformersTokenizer(Tokenizer):
 
 
 class TransformersEngine(Engine):
-    def __init__(self, 
-                 model, 
-                 tokenizer, 
-                 compute_log_probs: bool, 
-                 chat_template=None, 
-                 enable_backtrack=True, 
-                 enable_ff_tokens=True, 
-                 enable_monitoring=True, 
-                 **kwargs):
+    def __init__(
+        self,
+        model,
+        tokenizer,
+        compute_log_probs: bool,
+        chat_template=None,
+        enable_backtrack=True,
+        enable_ff_tokens=True,
+        enable_monitoring=True,
+        **kwargs,
+    ):
         # fill in default model value
         if model is None:
             model = os.environ.get("TRANSFORMERS_MODEL", None)
@@ -406,10 +405,19 @@ class TransformersEngine(Engine):
         self.model_obj = self._model(model, **kwargs)
 
         if not isinstance(model, str):
-            self.model = model.__class__.__name__
+            try:
+                self.model = self.model_obj.config["_name_or_path"]
+            except KeyError:
+                self.model = self.model_obj.__class__.__name__
+        else:
+            self.model = model
         self.device = self.model_obj.device  # otherwise note the current device
 
-        self._past_key_values: Union[transformers_package.Cache, tuple[tuple[torch.Tensor, ...], tuple[torch.Tensor, ...]], None] = None
+        self._past_key_values: Union[
+            transformers_package.Cache,
+            tuple[tuple[torch.Tensor, ...], tuple[torch.Tensor, ...]],
+            None,
+        ] = None
         self._cached_logits = None
         self._cached_token_ids: list[int] = []
 
@@ -499,7 +507,10 @@ class TransformersEngine(Engine):
             # TODO: this seems to get set to the length of the first sequence we pass for models using
             # StaticCache or HybridCache. We need to initialize our own cache with a large enough size
             # if we want to continue generation with the same cache.
-            if isinstance(past_key_values, (transformers_package.StaticCache, transformers_package.HybridCache)):
+            if isinstance(
+                past_key_values,
+                (transformers_package.StaticCache, transformers_package.HybridCache),
+            ):
                 # The __init__ API isn't consistent between different cache types, but there seems to be consistency
                 # between these two types, so we can use the same logic for both.
                 warnings.warn("Cache is too small. Re-initializing cache with larger size.")
@@ -508,17 +519,21 @@ class TransformersEngine(Engine):
                 device = self.model_obj.device
                 hf_device_map = getattr(self.model_obj, "hf_device_map", {})
                 # hf_device_map is not always a complete mapping of layers to devices...
-                layer_device_map = {k: hf_device_map.get(k, device) for k in range(config.num_hidden_layers)}
+                layer_device_map = {
+                    k: hf_device_map.get(k, device) for k in range(config.num_hidden_layers)
+                }
                 self._past_key_values = cache_type(
                     config=config,
                     batch_size=past_key_values.batch_size,
                     # Double the cache size to be safe
-                    max_cache_len=len(token_ids)*2,
+                    max_cache_len=len(token_ids) * 2,
                     dtype=past_key_values.dtype,
                     layer_device_map=layer_device_map,
                 )
             else:
-                warnings.warn(f"Cache is too small. Resetting cache (no method implemented to resize cache for type {type(past_key_values)}).")
+                warnings.warn(
+                    f"Cache is too small. Resetting cache (no method implemented to resize cache for type {type(past_key_values)})."
+                )
                 self._past_key_values = None
             past_length = 0
         elif past_length > num_cached:
@@ -531,7 +546,9 @@ class TransformersEngine(Engine):
                 if hasattr(past_key_values, "crop"):
                     self._past_key_values.crop(past_length)
                 else:
-                    warnings.warn(f"Cropping unsupported for cache type: {type(self._past_key_values)}. Resetting cache.")
+                    warnings.warn(
+                        f"Cropping unsupported for cache type: {type(self._past_key_values)}. Resetting cache."
+                    )
                     if hasattr(self._past_key_values, "reset"):
                         # Use built-in reset method if available to avoid constructing/allocating a new cache
                         self._past_key_values.reset()
@@ -593,7 +610,9 @@ class TransformersEngine(Engine):
 
         return self._cached_logits
 
-    def get_per_token_topk_probs(self, token_ids: list[int], top_k: int = 5) -> list[GenTokenExtra]:
+    def get_per_token_topk_probs(
+        self, token_ids: list[int], top_k: int = 5
+    ) -> list[GenTokenExtra]:
         tokenizer = self.tokenizer._orig_tokenizer
 
         # NOTE (loc) - assume batch size of 1
@@ -649,31 +668,28 @@ class TransformersEngine(Engine):
 
         return batch[0]
 
-
-class Transformers(Model):
-    def __init__(
+    def apply_chat_template(
         self,
-        model=None,
-        tokenizer=None,
-        echo=True,
-        compute_log_probs=False,
-        chat_template=None,
-        enable_backtrack=True,
-        enable_ff_tokens=True,
-        enable_monitoring=True,
-        **kwargs,
-    ):
-        """Build a new Transformers model object that represents a model in a given state."""
-        super().__init__(
-            TransformersEngine(
-                model,
-                tokenizer,
-                compute_log_probs,
-                chat_template=chat_template,
-                enable_backtrack=enable_backtrack,
-                enable_ff_tokens=enable_ff_tokens,
-                enable_monitoring=enable_monitoring,
-                **kwargs,
-            ),
-            echo=echo,
+        messages: Sequence[Message],
+        active_message: Message,
+        tools: Optional[list[Any]],
+    ) -> str:
+        # TODO: move onto the tokenizer
+        # This is a hack to get around the fact that apply_chat_template won't properly continue the final message
+        # if it is empty. We add a sentinel value to the final message, and then remove it after the fact.
+        sentinel_value = "<|FINAL_MESSAGE_SENTINEL_VALUE|>"
+        messages = [{"role": m["role"], "content": m["data"].get("content", "")} for m in messages]
+        active_message = {
+            "role": active_message["role"],
+            "content": active_message.get("content", "") + sentinel_value,
+        }
+        prompt = self.tokenizer._orig_tokenizer.apply_chat_template(
+            conversation=(list(messages) + [active_message]),
+            tools=tools,
+            chat_template=None,
+            continue_final_message=True,
+            add_generation_prompt=False,
+            tokenize=False,
         )
+        prompt = prompt[: prompt.rindex(sentinel_value)]
+        return prompt
