+++ tests/model_specific/test_visual.py
