+++ tests/unit/library/test_set_attribute.py
@@ -1,6 +1,7 @@
+import pytest
 from guidance import set_attribute, models, gen
 
-
+@pytest.mark.xfail(reason="set_attribute isn't currently implemented")
 def test_set_attribute():
     lm = models.Mock(b"<s>1234233234<s>")
     with set_attribute("echo", False):
