+++ tests/unit/library/test_silent.py
@@ -1,6 +1,7 @@
+import pytest
 from guidance import models, silent
 
-
+@pytest.mark.xfail(reason="silent isn't currently implemented")
 def test_basic():
     lm = models.Mock()
     lm += "Start text"
